+++ lib/src/package_analyzer.dart
@@ -3,7 +3,6 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:collection';
 import 'dart:convert';
 import 'dart:io';
 
@@ -14,7 +13,6 @@ import 'package:pub_semver/pub_semver.dart';
 
 import 'code_problem.dart';
 import 'download_utils.dart';
-import 'library_scanner.dart';
 import 'license.dart';
 import 'logging.dart';
 import 'maintenance.dart';
@@ -115,7 +113,6 @@ class PackageAnalyzer {
         packageVersion: null,
         pubspec: options.verbosity == Verbosity.compact ? null : pubspec,
         pkgResolution: null,
-        dartFiles: null,
         licenses: null,
         tags: null,
         report: null,
@@ -129,24 +126,7 @@ class PackageAnalyzer {
           '[report the issue](https://github.com/dart-lang/pana/issues).');
     }
 
-    final package = pubspec.name;
     final usesFlutter = pubspec.usesFlutter;
-
-    Set<String> unformattedFiles;
-    try {
-      unformattedFiles = SplayTreeSet<String>.from(
-          await _toolEnv.filesNeedingFormat(pkgDir, usesFlutter,
-              lineLength: options.lineLength));
-
-      assert(unformattedFiles.every((f) => dartFiles.contains(f)),
-          'dartfmt should only return Dart files');
-    } on ToolException catch (e) {
-      errors.add(messages.runningDartfmtFailed(usesFlutter, e.message));
-    } catch (e, stack) {
-      log.severe('`dartfmt` failed.\n$e', e, stack);
-      errors.add(messages.runningDartfmtFailed(usesFlutter, e.toString()));
-    }
-
     final upgrade = await _toolEnv.runUpgrade(pkgDir, usesFlutter);
 
     PkgResolution pkgResolution;
@@ -193,12 +173,6 @@ class PackageAnalyzer {
               '```\n$message\n```\n');
     }
 
-    Map<String, List<String>> allDirectLibs;
-    Map<String, List<String>> allTransitiveLibs;
-    Set<String> reachableLibs;
-
-    LibraryScanner libraryScanner;
-
     List<CodeProblem> analyzerItems;
 
     if (pkgResolution != null && options.dartdocOutputDir != null) {
@@ -221,40 +195,6 @@ class PackageAnalyzer {
 
     final tags = <String>[];
     if (pkgResolution != null) {
-      try {
-        var overrides = [
-          LibraryOverride.webSafeIO('package:http/http.dart'),
-          LibraryOverride.webSafeIO('package:http/browser_client.dart'),
-          LibraryOverride.webSafeIO(
-              'package:package_resolver/package_resolver.dart'),
-        ];
-
-        libraryScanner = LibraryScanner(_toolEnv.dartSdkDir, package, pkgDir,
-            overrides: overrides);
-        assert(libraryScanner.packageName == package);
-      } catch (e, stack) {
-        log.severe('Could not create LibraryScanner', e, stack);
-        errors.add('LibraryScanner creation failed: `$e`.');
-      }
-
-      if (libraryScanner != null) {
-        try {
-          log.info('Scanning direct dependencies...');
-          allDirectLibs = await libraryScanner.scanDirectLibs();
-        } catch (e, st) {
-          log.severe('Error scanning direct libraries', e, st);
-          errors.add('Error scanning direct libraries: `$e`.');
-        }
-        try {
-          log.info('Scanning transitive dependencies...');
-          allTransitiveLibs = await libraryScanner.scanTransitiveLibs();
-          reachableLibs = _reachableLibs(allTransitiveLibs);
-        } catch (e, st) {
-          log.severe('Error scanning transitive libraries', e, st);
-          errors.add('Error scanning transitive libraries: `$e`.');
-        }
-      }
-
       if (dartFiles.isNotEmpty) {
         try {
           analyzerItems = await _pkgAnalyze(pkgDir, usesFlutter, options);
@@ -277,53 +217,6 @@ class PackageAnalyzer {
         }
       }
     }
-    String pkgPlatformConflict;
-
-    final files = SplayTreeMap<String, DartFileSummary>();
-    for (var dartFile in dartFiles) {
-      final size = fileSize(pkgDir, dartFile);
-      if (size == null) {
-        log.warning('File deleted: $dartFile');
-      }
-      final isFormatted = unformattedFiles == null
-          ? null
-          : !unformattedFiles.contains(dartFile);
-      final fileAnalyzerItems =
-          analyzerItems?.where((item) => item.file == dartFile)?.toList();
-      final codeErrors =
-          fileAnalyzerItems?.where((cp) => cp.isError)?.toList() ?? const [];
-      final platformBlockers =
-          codeErrors.where((cp) => cp.isPlatformBlockingError).toList();
-      var uri = toPackageUri(package, dartFile);
-      final libPlatformBlocked = platformBlockers.isNotEmpty &&
-          (reachableLibs == null || reachableLibs.contains(uri));
-      var directLibs = allDirectLibs == null ? null : allDirectLibs[uri];
-      var transitiveLibs =
-          allTransitiveLibs == null ? null : allTransitiveLibs[uri];
-      if (libPlatformBlocked) {
-        pkgPlatformConflict ??=
-            'Error(s) in $dartFile: ${platformBlockers.first.description}';
-      }
-      files[dartFile] = DartFileSummary(
-        uri: uri,
-        size: size,
-        isFormatted: isFormatted,
-        codeProblems: fileAnalyzerItems,
-        directLibs: directLibs,
-        transitiveLibs:
-            options.verbosity == Verbosity.verbose ? transitiveLibs : null,
-      );
-    }
-
-    if (analyzerItems != null) {
-      final reportedFiles = analyzerItems.map((i) => i.file).toSet();
-      final knownFiles = files.values.map((f) => f.path).toSet();
-      final unattributedFiles = <String>{...reportedFiles}
-        ..removeAll(knownFiles);
-      if (unattributedFiles.isNotEmpty) {
-        log.warning('Unattributed files from dartanalyzer: $unattributedFiles');
-      }
-    }
 
     var licenses = await detectLicensesInDir(pkgDir);
     licenses = await updateLicenseUrls(
@@ -338,7 +231,6 @@ class PackageAnalyzer {
       pubspec: options.verbosity == Verbosity.compact ? null : pubspec,
       pkgResolution:
           options.verbosity == Verbosity.compact ? null : pkgResolution,
-      dartFiles: options.verbosity == Verbosity.compact ? null : files,
       licenses: licenses,
       tags: tags,
       report: await createReport(options, pkgDir, _toolEnv),
@@ -362,20 +254,6 @@ class PackageAnalyzer {
         .toList();
     list.sort();
     return list;
-  }
-
-  Set<String> _reachableLibs(Map<String, List<String>> allTransitiveLibs) {
-    final reached = <String>{};
-    for (var lib in allTransitiveLibs.keys) {
-      if (lib.startsWith('package:')) {
-        final path = toRelativePath(lib);
-        if (path.startsWith('lib/') && !path.startsWith('lib/src')) {
-          reached.add(lib);
-          reached.addAll(allTransitiveLibs[lib]);
-        }
-      }
-    }
-    return reached.intersection(allTransitiveLibs.keys.toSet());
   }
 }